Match Tor Browser and (new) Orfox UAS together
The two user agent strings are now nearly identical, we should try
combining the patterns (again). There are also new unit test strings
for this.

This roughly combines the pattern matching introduced in
749fd988c32e764e7fc8dde1f59ab7f06a63f0cd and
644e1c21b682a911acbea567f55d156a084a8c5f.

We can remove the old Orfox-specific pattern (Gecko/38) after
users have a few weeks for upgrading to the new version.